fix(v2.2 D): smalls sweep + bit-exact sync gate timestamp
M2: dedup-origin comments now match DedupOrigin's actual bank case.
L1: DateFieldParser dot-separator month-first ternary fixed.
L2: dedup merge repoints duplicateOfID + linkedTransactionID to keeper.
L3: anchor-only reconciliation records + shows unabsorbed residual
(carried through backup DTOs); math unchanged.
L5: BNR rate parsed and persisted as exact Decimal; all display
multiplies consume rateDecimal (no Double->Decimal noise).
CI fix: BankSyncGate persists timeIntervalSinceReferenceDate (bit-exact
round-trip, key renamed) - closes the run-33258846602 failure.
Follow-ups: Raport next-payment wired to schedule stamps; backup DTO
carries scheduleIndex; below-interest + truncated-cost strings wired.

+++ Bani/People/ReceivablesRollup.swift
@@ -38,17 +38,20 @@ struct ReceivablesSummary: Equatable, Sendable {
 /// recurring receivable once, never once per past/future occurrence.
 enum ReceivablesRollup {
 
-    static func ronValue(_ item: ScheduledItemSnapshot, rate: Double?) -> Decimal? {
+    /// L5: `rate` is a `Decimal` (see `RateService.rateDecimal`) so this never
+    /// re-derives a `Decimal` from the display `Double` — the source of the FX
+    /// display noise this fix removes.
+    static func ronValue(_ item: ScheduledItemSnapshot, rate: Decimal?) -> Decimal? {
         switch item.currency {
         case .ron: return item.amount
-        case .eur: return rate.map { item.amount * Decimal($0) }
+        case .eur: return rate.map { item.amount * $0 }
         }
     }
 
     /// Groups pending/incoming items by counterparty (blank counterparties
     /// excluded — there is no "who" to owe). Done items and outgoing items
     /// (money the client owes, not receivables) are excluded entirely.
-    static func build(_ items: [ScheduledItemSnapshot], rate: Double?) -> ReceivablesSummary {
+    static func build(_ items: [ScheduledItemSnapshot], rate: Decimal?) -> ReceivablesSummary {
         var byName: [String: [(item: ScheduledItemSnapshot, ronValue: Decimal)]] = [:]
         var order: [String] = []
